countermeasures against cisco dhcp snooping.
make the dhcp message type invalid/unknown so dhcp snooping doesnt
assume it knows what's going on and filter the message when it's
out of order or something.

allow the dhcp hardware address to be specified separately to the
mac address we're trying to wol. this allows us to flood the packet
even if the switch thinks it knows where the address is.